requirements.txt
youtube_transcript_api
pytube
streamlit
google-generativeai
python-dotenv
pathlib